Trait sp_lightclient::Storage

pub trait Storage<Header: HeaderT> {
    fn record_size(&self) -> u32;
    fn segment_size(&self) -> u32;
    fn header(
        &self,
        query: <Header as HeaderT>::Hash
    ) -> Option<HeaderExt<Header>>;
    fn store_header(
        &mut self,
        header_ext: HeaderExt<Header>,
        as_best_header: bool
    );
    fn best_header(&self) -> HeaderExt<Header>;
}

Storage responsible for storing headers

Required Methods

fn record_size(&self) -> u32

Record size

fn segment_size(&self) -> u32

Segment size

fn header(&self, query: <Header as HeaderT>::Hash) -> Option<HeaderExt<Header>>

Queries a header at a specific block number or block hash

fn store_header(&mut self, header_ext: HeaderExt<Header>, as_best_header: bool)

Stores the extended header. as_best_header signifies of the header we are importing is considered best

fn best_header(&self) -> HeaderExt<Header>

Returns the best known tip of the chain