Crate sp_lightclient

Light client substrate primitives for Subspace.

Structs

ChainConstants

Chain constants.

HeaderBelowArchivingDepthError

Error type that holds the current finalized number and the header number we are trying to import.

HeaderExt

HeaderExt describes an extended block chain header at a specific height along with some computed values.

HeaderImporter

Verifies and import headers.

NextDigestItems

Type to hold next digest items present in parent header that are used to verify the immediate descendant.

Enums

ImportError

Error during the header import.

StorageBound

Defines the storage bound for the light client store.

Traits

Storage

Storage responsible for storing headers.