Trait sp_lightclient::Storage

pub trait Storage<Header: HeaderT> {
    // Required methods
    fn chain_constants(&self) -> ChainConstants<Header>;
    fn header(
        &self,
        hash: <Header as HeaderT>::Hash
    ) -> Option<HeaderExt<Header>>;
    fn store_header(
        &mut self,
        header_ext: HeaderExt<Header>,
        as_best_header: bool
    );
    fn best_header(&self) -> HeaderExt<Header>;
    fn headers_at_number(
        &self,
        number: <Header as HeaderT>::Number
    ) -> Vec<HeaderExt<Header>>;
    fn prune_header(&mut self, hash: <Header as HeaderT>::Hash);
    fn finalize_header(&mut self, hash: <Header as HeaderT>::Hash);
    fn finalized_header(&self) -> HeaderExt<Header>;
    fn store_segment_commitments(
        &mut self,
        segment_commitments: BTreeMap<SegmentIndex, SegmentCommitment>
    );
    fn segment_commitment(
        &self,
        segment_index: SegmentIndex
    ) -> Option<SegmentCommitment>;
    fn number_of_segments(&self) -> u64;
    fn max_pieces_in_sector(&self) -> u16;
}

Storage responsible for storing headers.

Required Methods

fn chain_constants(&self) -> ChainConstants<Header>

Returns the chain constants.

fn header(&self, hash: <Header as HeaderT>::Hash) -> Option<HeaderExt<Header>>

Queries a header at a specific block number or block hash.

fn store_header(&mut self, header_ext: HeaderExt<Header>, as_best_header: bool)

Stores the extended header. as_best_header signifies of the header we are importing is considered best.

fn best_header(&self) -> HeaderExt<Header>

Returns the best known tip of the chain.

fn headers_at_number( &self, number: <Header as HeaderT>::Number ) -> Vec<HeaderExt<Header>>

Returns headers at a given number.

fn prune_header(&mut self, hash: <Header as HeaderT>::Hash)

Prunes header with hash.

fn finalize_header(&mut self, hash: <Header as HeaderT>::Hash)

Marks a given header with hash as finalized.

fn finalized_header(&self) -> HeaderExt<Header>

Returns the latest finalized header.

fn store_segment_commitments( &mut self, segment_commitments: BTreeMap<SegmentIndex, SegmentCommitment> )

Stores segment commitments for fast retrieval by segment index at or below finalized header.

fn segment_commitment( &self, segment_index: SegmentIndex ) -> Option<SegmentCommitment>

Returns a segment commitment for a given segment index.

fn number_of_segments(&self) -> u64

Returns the stored segment count.

fn max_pieces_in_sector(&self) -> u16

How many pieces one sector is supposed to contain (max)